+++ open_alchemy/column_factory/__init__.py
@@ -31,7 +31,6 @@ def column_factory(
     schemas: oa_types.Schemas,
     required: typing.Optional[bool] = None,
     logical_name: str,
-    model_name: str,
     model_schema: oa_types.Schema,
 ) -> typing.Tuple[types.TReturnValue, _TReturnSchema]:
     """
@@ -42,7 +41,6 @@ def column_factory(
         schemas: Used to resolve any $ref.
         required: Whether the object property is required.
         logical_name: The logical name in the specification for the schema.
-        model_schema: The name for the model.
         model_schema: The schema for the model.
 
     Returns:
@@ -72,14 +70,12 @@ def column_factory(
             schemas=schemas,
             required=required,
             logical_name=logical_name,
-            model_name=model_name,
             model_schema=model_schema,
         )
 
     # Handle arrays
     return array_ref.handle_array(
         schema=schema,
-        model_name=model_name,
         model_schema=model_schema,
         schemas=schemas,
         logical_name=logical_name,