+++ tests/open_alchemy/test_model_factory.py
@@ -111,7 +111,6 @@ def test_single_property_required_missing(mocked_column_factory: mock.MagicMock)
         schemas=schemas,
         logical_name="id",
         required=None,
-        model_name=model_name,
         model_schema=model_schema,
     )
 
@@ -141,7 +140,6 @@ def test_single_property_not_required(mocked_column_factory: mock.MagicMock):
         schemas=schemas,
         logical_name="id",
         required=False,
-        model_name=model_name,
         model_schema=model_schema,
     )
 
@@ -171,7 +169,6 @@ def test_single_property_required(mocked_column_factory: mock.MagicMock):
         schemas=schemas,
         logical_name="id",
         required=True,
-        model_name=model_name,
         model_schema=model_schema,
     )