+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## Version _next_
+
+- Add schemas pre-processor that extracts the required back references.
+
 ## Version 1.3.0 - 2020-07-12
 
 - Add support for generic JSON data for properties.